third_party/WebKit/LayoutTests/crypto/hmac-generate-key-expected.txt

Test generating a HMAC key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Generating a key with default length...
error is: TypeError: Failed to execute 'generateKey' on 'SubtleCrypto': parameter 1 ('algorithm') is not an object.
error is: SyntaxError: HmacKeyGenParams: hash: Missing or not a dictionary
error is: SyntaxError: HmacKeyGenParams: length: Is not a number
error is: SyntaxError: HmacKeyGenParams: length: Is not a number
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'HMAC'
PASS key.algorithm.length is 512
PASS key.usages is ["sign", "verify"]

Generating a key with custom length...
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'HMAC'
PASS key.algorithm.length is 40
PASS key.usages is ["sign"]
PASS successfullyParsed is true

TEST COMPLETE
